feat(audio): Add cartwall audio processing and routing
Introduces a dedicated audio destination for the cartwall stream. This allows the cartwall audio to be processed alongside other audio sources and routed to the main bus.

Updates the `RemoteStudio` component to include the cartwall stream in the combined audio stream when going on air.

Also, the `AudioMixer` component is updated to enforce that the cartwall stream is always sent to the main bus when in studio playout mode, and this send option is disabled for cartwall source.